Fix calculation bugs and add validation
Fixes:
- Add missing file_utils import (get_output_filename, ensure_directory_exists)
- Change quantity/stock from int to float for fractional share support
- Add null check for ForeignTax with clear error message
- Round interest condensation to 4 decimals to avoid float precision issues
- Add validation that trade close date >= open date

Tests:
- Add fractional shares tests (3 tests)
- Add 4-decimal rounding test for interest condensation

+++ gains.py
@@ -6,9 +6,9 @@ from typing import Literal
 @dataclass
 class KDVPSecurity:
     date: datetime.date
-    quantity: int
+    quantity: float
     value: float
-    stock: int
+    stock: float
 
 
 @dataclass
